[WifiMacRandomizationTest] Fix flaky issue
test_roaming_mac_randomization is flaky as the roaming to the mirror AP
does not always happen. Since this test verifies the randomized MAC
address after roaming and does not test roaming code itself, we are
setting the the attenuation to 95dbm instead of 60dbm. A dictionary
specifying the attenuation is required in the config file.

Bug: 145539505
Test: Verified the changes

Change-Id: I1c9a59b45fbcc6c5540ecd3b520ab42a14821f50