Adds version of the rock_bottom tests that relies on the older APK.
Adds an example test on how to use the Power.apk version.
In case power-tests.apk processes get killed/frozen by OEM devices, having
the older APK as a backup is an option. It is still not confirmed but
at least one instance of the Activity Manager force stopping it was found
while the test was still running:

```
04-28 23:26:39.871  1000  1443  1487 I ActivityManager: Force stopping com.google.android.device.power appid=10300 user=-1: deletePackageX
04-28 23:26:39.873  1000  1443  1487 I ActivityManager: Killing 30318:com.google.android.device.power/u0a300 (adj 0): stop com.google.android.device.power,deletePackageX
04-28 23:26:39.876  1000  1443  1487 W ActivityManager: Crash of app com.google.android.device.power running instrumentation ComponentInfo{com.google.android.device.power/androidx.test.runner.AndroidJUnitRunner}
04-28 23:26:39.876  1000  1443  1487 I ActivityManager: Force stopping com.google.android.device.power appid=-1 user=0: finished inst
```

Bug: 154227079
Test: Tested on apollo-1
Change-Id: I8c1fb1f9078e04310454e143c32f096126a354ad