Removes automanage_usb feature.
Since bits doesn't allow to add markers (logs) to closed collections
but the markers are defined by events that happen on the device, it
doesn't make much sense to tandem stop_collection with connect_usb.

usb connection is needed to reach the device for getting the markers
information, and stop_collection makes so that you can no longer
add markers to a collection.

go/acts-bits
Bug: 155889181
Test: Unit tests passing

Change-Id: I736414f229e395d0af493f9d0e42c0f64d2ef71f