Add configuration to run test through policy
Add an optional flag to the associate method used by WLAN tests so that
we can optionally test through the WLAN policy path. There is also an
added target security field since the policy layer needs to know the
security type of the network to connect to.

The WlanSecurityComplianceABGTest is modified so that if the
association_mechanism field is set to policy in the configuration file,
the test will associate throuh the policy layer, and if not it will test
the core/drivers layer.

Bug: None
Test: act.py -c <path to config> -tc WlanSecurityComplianceABGTest
Change-Id: I8319f16f81d8b636003f0b4e91d7efcb3d658c89