wifi: finetune connect flow for keypad method
The keypad method requires the sender to send pins to the receiver.
If the receiver clicks confirm button quickly, it might miss the
returning negotiation request.

In a real use case, an user needs some time to input pins, as a result,
the confirm button should be pressed after few seconds.

Bug: 163532173
Test: act.py
      -c tools/test/connectivity/acts/tests/google/wifi/p2p/config/wifi_p2p.json \
      -tb WifiP2pAllAttached \
      -tc WifiP2pManagerTest
Change-Id: Iabec1a5b98b828b556e5a8faa51f864b9178c2d3