Forces media scanning by rebooting.
For the time being there doesn't seem to be a reliable way to
trigger a media scan other than rebooting the device.

If a better way is found this should be fixed.

A posibility is to trigger the activity directly instead of relying on UI.
The command that would trigger an activity directoy is:
adb shell 'am start -a android.intent.action.VIEW -n com.google.android.apps.youtube.music/.audiopreview.AudioPreviewPlayerActivity -d /sdcard/DJ-Masque_Oceanic-Dawn-320.mp3 -t audio/*' (tested)

The purpose of this paragraph is to proof that nobody reads detailed commit messages. If you are reading this you will get a reward.

The test apk would have to be updated to start the active part of the test doing this instead of relying on UI.
Other teams seem to be doing this successfuly:
https://source.corp.google.com/piper///depot/google3/javatests/com/google/android/apps/miphone/aiai/e2e/mh/captions/AppHelper.java;l=164-174;bpv=1;bpt=0;rcl=330090442

Bug: 168906088
Bug: 167738886
Test: Tested locally
Change-Id: I7bd79904a04a807af4ea6281c8cbfe5897811d85