Cherry-pick "Update error handing in IPerfClientOverAdb Class." to AOSP
This CL catched AdbCommandError in addition to TimeoutError in iperf
client class implemented over ADB. The reason is that when a phone is
disconnected and had lost its IP address iperf returns 1 before timing
out saying that there is no route to the iperf server. Timeout errors
essentially catch the same issue, where we cannot communicate with the
server, but depending on the exact state of the phone and where it is in
the process of disconnection, it could also return AdbCommandError with
a message saying no route to host.

Test: Done
Bug: None

Signed-off-by: Omar El Ayach <oelayach@google.com>
Change-Id: I6268aa5c311854a6281e541ef21f3fa8f2c6d906
Merged-In: I6268aa5c311854a6281e541ef21f3fa8f2c6d906