Add fd setup and teardown commands from config
Allows setup and teardown commands to be passed in via the ACTS config to be run on a
Fuchsia device by its ontroller. This is especially useful for running commands
specific to hardware or build type.

Also, adds a exception catch to SoftApTest when an interface is expected to be deleted.

Bug: None
Test: Ran SoftApTests, which had cast_agent.cmx interfering with policy layer before.
Change-Id: I3d8ede0e61a996e1da14dd161633a77b7e3ab835