Updated iperf client to take fuchsia name plus clean up
Updated the iperf client controller to be able to take a fuchsia
device hostname and resolve it to a real ip address. Plus
cleaned up Fuchsia device to not call get_fuchsia_mdns_ipv6_address
to limit look up times.  Also added a ping in the
get_fuchsia_mdns_ipv6_address so the caller can know the device
is resolved and alive.

Bug: None
Test: Ran RvR test suite that runs iperf between 2 clients.
Change-Id: Ic6df961da5042faa329d64872461d03306db7fff