Add BSS scan before connect for core associations
BSS scan and selection was moved to the policy layer, meaning for
core connections, the tests need to scan for and select a BSS
to pass to the connect call.

Small edits to WlanScanTest, which was written before the wlan_device
abstract device was created, and PolicyScanTest, which was using
the old core connect call.

Bug: None
Test: Ran affected tests.
Change-Id: I972806b4f80254998b31726c982071709a18f75b