Initialize WlanController for FuchsiaDevice
In the spirit of reducing FuchsiaDevice complexity, this change
intializes a WlanController library, to be imported by FuchsiaDevice,
which will contain WLAN core functions (similar to WlanPolicyController
for policy layer).

In this lib (to start), are some wlan iface query helper functions,
as well as placeholder configure and deconfigure methods to match the
policy controller.

Bug: None
Test: Used in relation chain
Change-Id: I6cba4172baebff497ed575e2f521a274f45c07d4