Automatically kill zombie iperf servers blocking intended iperf port.
This CL improves robustness of iperf servers over ssh by checking if
there are existing iperf servers blocking the port we intend to start a
new iperf server on. If a server is found, it is killed. Previously, the
controller would attempt to start a controller whether or not a port was
occupied without raising any warnings or errors  leading to garbage
(empty) iperf results files when the port was infact not available.

Test: Done
Bug: None

Signed-off-by: Omar El Ayach <oelayach@google.com>
Change-Id: I0735228f7fabaf454918bee322cc3b2190e20fb1