Fix policy _configure_wlan retry loops
Fuchsia policy client controllers may be created successfully, but then have their
channels closed later (due to asynchronicity). If this occurs, the error will manifest
the next time the controller is used.

The updated retry loop creates a client controller and attempts to use it. If either
fail, then both need to be retried.

Bug: 71244
Test: Ran policy tests
Change-Id: I769bd6039acd7ddd2a6eb13ac7d392a185d5802d