ACTS Wifi Autojoin Tests: exercise Make Before Break
Previously, we don't assert that Wifi is connected
at the beginning of each test. Thus, it's possible
that Wifi is going from disconnected to connected.

Ensure that at the beginning of each test case, we
are already connected to a network. Then switching
to another network will guarantee a Wifi-to-Wifi
switch, potentially exercising MBB if the conditions
are right. Note that we never explicitly check for
the occurrence of 2 concurrent STAs. We are merely
ensuring that Wifi works correctly after Wifi-to-
Wifi switch regardless of MBB or BBM.

Bug: 179400219
Test: Will be tested continuously for BBM
Test: Will file a test request for MBB
Change-Id: I41fd5596b24c1ec43825128ee6fa7c416f4d591b