[WifiNewSetupWifiToWifiAutoJoinTest] Test case fixes
Add fix for test case name.
Add UUIDs for the new tests.

Presubmit and regression tests are failing with b/183282252 as the test
case decorator are inherited in the new test class. Testtracker UUIDs
are unique and apply to only 1 test case. ACTS as a result throws
exception due to this mismatch. Inherited python decorators cannot be
overwritten.

Bug: 183282252
Bug: 182936315
Fixes: 183282252
Fixes: 182936315
Change-Id: If0017d836abed02931db6663f132346baf402940