Add bug report on fail flag to FuchsiaDevices
Added logic to take a bug report when a test fails on Fuchsia, only
if the "take_bug_report_on_fail" flag is set. If flag is not present
on DUT (e.g Android), a bug report will be taken anyway.

This allows Fuchsia devices to opt-in or out of taking snapshots for each failed test.

Bug: 72522
Test: Ran tests with both Fuchsia and Android
Change-Id: I7842e5e4e35d1ddacf9bbaa63fa5411957b63223