Stop generation of individual metrics for blackbox logger.
This change eliminates the creation of individual proto messages files
per metric results. All metrics are bundled into collection of messages.

Some file indexes have a limited number of files they can handle, the
creation of individual messages per metrics pushed some tools over
the limits.

A second bug was discovered in the process of fixing this. If
two loggers of the same time  were used concurrently one would override
the results generated by the other. Currently no single test is using
concurrent loggers of the same type, but this is technically allowed
it should be supported. The fix for was to give unique names to
result files by using a hash function.

Bug: 182320010
Bug: 171213924
Test: Unit tests updated and expanded
Change-Id: I09c1a60addce3eef60fdbc3eb907a887245f3eb3