Refactor SoftApTest
The SoftApTest suite was written extrememly quickly to meet urgent
testing needs. As such, it contains some sloppy repeated code and
is confusing to understand and edit.

I filed two bugs to track the work to refactor it and make it
a bit cleaner (adding here since gerrit doesn't play well with monorail):

fxb/59337: Cleanup SoftApTests
fxb/59306: Fix simultaneous Client/Ap parallel streaming

Both are resolved with this change, as I removed repeated code and
unnecessarily complex logic, and the iperf flakiness was due
to a previously fixed issue.

Bug: None
Test: Ran all SoftApTests to verify they work still.

Change-Id: Ibe345edab598b8828ec70afa92b96c4ed47f2e88