Report version of Fuchsia DUT after flashing
Updates Sponge properties with the DUT's version of Fuchsia after
FlashTest. The DUT is determined from a list of devices by looking for
the "device_type" property. If this property is not available, the first
device specified will be considered as the DUT.

In cases when flashing is not desirable but version reporting is
necessary, run ACTS with "FlashTest:test_report_dut_version".

Change-Id: Ib2f3eae55e942e8c439fa061987d15d5df4570b7