Changed the bridge name of the ap and updated some RvR code
There is a race conidtion where, on the AP, if an interface
is brought up and not on the block list by shill, shill will
take over control of the interface and what IP addresses can
be applied to this interface.  In the case of bridge testing
with a router advertisement server on the AP, the interface
has a brief time where it can connect before shill takes over
but once shill takes over the router advertisment server
can no longer send out router advertisment messages.  By
default in whirlwind there are some interfaces that are
already configured for this purpose so one of them is being
reused in this change to make sure that interface is not
taken over by shill.

Also cleaning up a chunk of code that would never be hit
in the RvR test suite.

Bug: None
Test: Ran the RvR IPv6 test suite.
Change-Id: Id7f6113a113865e55b4925a7656e1a734eb476e5