Add resilience to Fuchsia mDNS lookup
Allow mDNS socket binding errors relating to not being able to find the
adapter for an IP address. This may occur when a large number of devices
are connected to the host, with muliple tests running parallel. When a
device connected to a USB ethernet adapter goes offline, the
corresponding interface on the host is deleted. This is acceptable since
the device that went offline isn't guaranteed to be the device we're
searching for.

Fixes https://fxbug.dev/83980.

Bug: None
Test: Manual testing of mDNS discovery on Linux
Test: `act.py [...] -tc Sl4fSanityTest` on Linux and Mac

Change-Id: Ie174fbcc31f993abc2eeb3f13e7afa9d77ed7089