[fuchsia][wlan] Fix wlan_lib is_connected method.
A recent change to in-tree Fuchsia data structures
( https://fxrev.dev/554125 ) caused the Fuchsia wlan_lib
is_connected method to fail in all cases. This change corrects the
error. Affected Fuchsia ACTS tests include BeaconLossTest, but also
probably WlanRebootTest and WlanRvrTest. This fix was verified on
BeaconLossTest, but not on the other tests (which require specialized
hardware to run).

Fixes: https://fxbug.dev/84355
Test: act.py -c <config> -tc BeaconLossTest
Change-Id: Ic634b4a22579548529defdddc2de73f06f29d433