Wait for DAD to resolve before testing
This change fixes an error only seen with IPv6 RvR runs where the iperf
server and/or DUT isn't assigned a private-local IPv6 address by the
time the test checks. This was due to Duplicate Address Detection (DAD)
not completing in time after the test interface was cycled.

When the iperf server's interface is taken down then up, the IPv6
address assigned by radvr is lost and needs to be reassigned. Part of
the reassignment process is ensuring the IPv6 address is not taken by
another node in the network, a process coined as Duplicate Address
Detection and specified in RFC 4862. DAD can take, on average, around 5
seconds. This change adds polling for the yet-to-be-assigned IPv6
address up to a maximum timeout.

Bug: None
Test: act.py [...] -tc WlanRvrTest
Change-Id: Iab8d27a07f7d3c9f4ee430fd3837fffc7ecc6556