[fuchsia][wlan] Operating class and SoftAP regression tests.
Debugging of fxbug.dev/84777 has uncovered two possible regressions
involving channels in operating class 115 and operating class 124. This
change adds specific regression tests relevant to the ongoing debugging.
These regression tests currently fail, and TODOs in the change note
this state. Because the SoftAP was previously involved in
fxbug.dev/64279, this change also adds the SoftAP into the mix and
verifies that the client and SoftAP switch to the correct channel.

The wlan_lib.wlanStatus function now takes an optional interface ID,
which is necessary to see the channel of the SoftAP.

Fuchsia bug: https://fxbug.dev/84777

Bug: None
Fixes: https://fxbug.dev/84701
Test: act.py -c <config> -tc ChannelSwitchTest
Change-Id: Ib106edea915e0c5440e2c40eb8b81c53d8ffd886