[fuchsia][acts] Fix BtFuchsiaEPTest
There were several issues with the BtFuchsiaEPTest that this CL
resolves:
1. The bt-avrcp component has been migrated to CFv2, which is able to
lazily load a component when a capability it offers is consumed. This
means we no longer need to explicitly start it before calling
accept_pairing. Instead, we just initialize avdtp on both devices.
2. BR/EDR semantics were changed such that no service connection can be
made without also pairing. Therefore, we force the devices to pair prior
to checking their status.

Bug: b/195343910
Change-Id: I7ad91aa7bff550af1f31b89d864d1df5bd7775ed