Add snapshots after fail on test
This change switches all WLAN tests to use AbstractDeviceWlanDeviceBaseTest class to leverage the on_fail method and take a snapshot of the device on failure. There is also some light refactoring and cleanup of unused and unnecessary imports and methods related to BaseTestClass and on_fail.

Fixes https://fxbug.dev/89573

Bug: None
Test: act.py [...] -tc
  -WlanDeprecatedConfigurationTest
  -WlanFacadeTest
  -WlanStatusTest
  -ConnectionStressTest
  -DownloadStressTest
  -PingStressTest
  -WlanRebootTest
  -WlanScanTest
  -WlanTargetSecurityTest
  -WlanInterfaceTest
  -ChannelSweepTest

Change-Id: Id36b2c198e49bb6cc1e2ee406eb65a51cbe37bb0