Isolate ffx to a single test-run environment
Updates ffx configuration to create files in a place that is specific to
the test run. This fixes multi-user permissions errors and the dreaded
random test flaps. Infrastructure-parallel tests may run entire test
suites that end up with overlapping daemons, causing the ffx invocations
to “upgrade” one daemon to another, which appears as a flap/restart to
another test.

Fixed: fxbug.dev/92243
Bug: None
Test: act.py [...] WlanSecurityComplianceABGTest:test_associate_11bg_sec_wpa2_psk_ptk_ccmp
Change-Id: I7956d676ee6ba7144ba106bf32e4f5e017d931e3