Stop accessing the internet while testing.
This changes the access_point controller to look for "wan_interface"
explicitly defined in the acts_config before falling back on pinging the
internet to determine which interface is setup for wan. This will break
some of the WLAN tests that currently rely on the internet and will
require a change to fix them. We're okay with this because 2 of the
tests are not critical and the third we will fix in a follow up change.

Change-Id: I43bd4038890880afad36ba916d8a181c139dc873