Catch StopIteration to comply with PEP 479
Fixes a bug where class teardown was failing with an uncaught
StopIteration exception. This behavior was introduced by a new change in
Python 3.7:

"PEP 479 is enabled for all code in Python 3.7, meaning that
StopIteration exceptions raised directly or indirectly in coroutines and
generators are transformed into RuntimeError exceptions. (Contributed by
Yury Selivanov in bpo-32670.)"
https://docs.python.org/3/whatsnew/3.7.html#changes-in-python-behavior

Bug: None
Test: local
Change-Id: Ibb10f3ae29c5e54504c2220949325ff3795da960