Update Mini-circuits attenuator controller.
This CL is a small update to the minicircuits attenuator controllers to
make sure implementation matches its docstring when the strict flag is
set to False. Namely, when set_atten is called with strinct=False, out
of bounds values should be set to 0 or max_atten whereas the earlier
implementation was sending values as is and the attenuator returning an
error code.

Test: Done
Bug: None

Signed-off-by: Omar El Ayach <oelayach@google.com>
Change-Id: I57d06d8de9943d355456f94a3e1f5b15dd22fccd
Merged-In: I57d06d8de9943d355456f94a3e1f5b15dd22fccd