Save space when flashing Fuchsia and abort on fail
Properly clean up the temporary directory used for storing the files for
flashing the device. Refactor to use tempfile and the `with` statement
for more reliability. Decompose flash into reboot_to_bootloader and
run_flash_script for clarity and brevity with the newly added `with`
statement.

Avoid copying the tar.gz file if it exists on disk, saving 700MB of
space for each active flash.

Abort all tests when FlashTest fails to flash the device.

Bug: None
Test: locally
Change-Id: Ibe53e644d058a7faa3ce0039cc8336f5384762cb