Defer init of ffx until needed
Refactor the FuchsiaDevice and FFX controller to defer initialization of
ffx until needed, which is when the ffx tool is first used.

Allow FlashTest to run without requiring a compatible verson of ffx
installed. This eliminates the failures we receive every time a Fuchsia
device upgrades version.

Speed up the setup for test suites that do not use ffx, which are
currently most Fuchsia test suites.

Fixes https://fxbug.dev/94504

Bug: None
Test: act.py [...] -tc FlashTest HiddenNetworksTest
Change-Id: I0bff9d31740df14505e5f4adafb9bf7333daef12