Statically set the version of protobuf
Fix the version of protobuf to stay 3.20.1 since this package is unlike
the other pip packages. When protobuf upgrades, all generated source
files needs to be regenerated. When a new feature of protobuf is
required, developers should be required to bump the version in setup.py
then regenerate all relevant files.

Fix: 234174079
Change-Id: I10a93175d626e49b664be4275b1f1fdf2970b0a1