Replace control_daemon with "run -d"
Replaces the Fuchsia control_daemon function with more lightweight
equivalents that don't require a persistent SSH connection. This is
done in an effort to simplify the device lifecycle by removing the need
to close the persistent SSH connection.

The replacement is running "run -d" over SSH to the Fuchsia device,
which hasn't failed on me yet.

Remove some of the calls to stop_sl4f_on_fuchsia_device in the reboot
method since they are no longer needed; a reboot of Fuchsia will stop
any running instance of SL4F. This was necessary before to close the
persistent SSH connection.

Bug: 230890623
Test: act.py [...] -tc FlashTest Sl4fSanityTest
Change-Id: I7a02533c04bb667f3704656a9480ac13bcdc3002