BSS Transition Management request test.
This test verifies that a DUT without BSS Transition Management
support does not roam upon receiving a BTM request. When ACTS can
see (or control) BTM support on a DUT, the test introduced in this
change can be trivially changed to test that BTM works correctly
on a device that has BTM support enabled and configured.

https://fxbug.dev/103440

Bug: None
Test: Manually verified this ACTS test works on a device.
Change-Id: I83deedb81acde52441a838fd37ccd59982a66436