Create SL4F module and refactor starting fuchsia_device services.
This change creates a new Sl4F module very similar to our FFX module. It
also refactors the fuchsia_device controller to do a lazy initialization of SL4F. This will help stabalize our FlashTest which is currently susceptible to issues with FFX compatibilty. This change also removes syslog streaming as it was deemed uneccessary for debugging and added extra complexity.
We also remove the id functionality attached to the SL4F commands
because the id was also unused by the SL4F facades. There is also more
cleanup in fuchsia_device removing unused class
FuchsiaDeviceLoggerAdapter. We moved start/stop_v1_component to ssh
module and logic for check_process_state there as well. We also removed
the setup and teardown commands that were run. This is a breaking change
for anyone using a Fuchsia build that does not kill
base_manager/session_manager.

Fix: 241310375

Test: act.py [...] -tc FlashTest, WlanRebootTest, HiddenNetworksTest,
SavedNetworksTest
Bug: 241310375

Change-Id: I208a8a167afda888f34c8956bdb223d0b94d849c