Remove dependency on requests for SL4F client
Remove all reference to the requests library for all of Fuchsia. Use of
request has been replaced by the standard library module "urllib".

Wait for the SL4F server to start accepting connections during the
initialization of the client. Removes race conditions where the SL4F
server is not ready for requests yet but the client is sending them
anyway.

Skip the ping check before every call to SL4F since we already verified
the connection to the SL4F server. Keep the ping check when a
TimeoutError has been caught however to assist debugging.

Move the function for waiting for an available port to a common utils
file for all Fuchsia modules.

Update Sl4fSanityTest to actually create a SL4F client and send a
request. This was previously broken by the change to lazy-initialize the
SL4F client.

Bug: 247016808
Test: act.py [...] -ci 3 -tc FlashTest Sl4fSanityTest \
  WlanRebootTest:test_soft_reboot_dut_2g_open_ipv4 \
  WlanRebootTest:test_hard_reboot_dut_5g_open_ipv6 \
Change-Id: If5cda4e26a4db099cb66791b5cddf75347d26a45