Major update to wifi_retail_ap classes.
This CL contains a major update to wifi_retail_ap, BlockingBrowser, and
the implementation of multiple Netgear APs included in the folder. The
update removed the dependence on the splinter python package, updates
implementation to support/require selenium==4.0.0 or later, and more
recent versions of webdriver-manager. This CL fixes several bugs in
NetgearRAXE500 and hardens implementation (which earlier required a
very specific and old version of splinter and selenium). This CL breaks
the implementation Netgear R7000, R7500, R8000, RAX80 and RAX120. A
warning has been added to these files and they have not been deleted,
rather maintained as a future TODO.

Test: Done locally on RAXE500
Bug: None

Signed-off-by: Omar El Ayach <oelayach@google.com>
Change-Id: I8443bb7a6945b88bd96c5e84b67b85292dcc0dfe